fix: Fix parseInt to parseFloat

+++ Public/JS/findWH.js
@@ -99,7 +99,7 @@ $(function () {
     $('#btn').click(function () {
         let warehouseID = $('#whID').text();
         var strPrice = $('#whPrice').text();
-        let price = parseInt(strPrice.substr(0, strPrice.length-2));
+        let price = parseFloat(strPrice.substr(0, strPrice.length-2));
         let useableArea = parseInt($('#whUseableArea').text());
         let wantArea = parseInt($('#area').val());
         let startDate = $('#startDate').text();
